examples: add phoenix-arduino-joystick
Create another example which should enumerate as a USB HID Joystick on the
attached computer and send the axes and button states in a report message.

This example works only on Arduino devices with HID support. These devices
use the main processor to access USB instead of a separate processor, which
allows them to provide additional USB endpoints that can be accessed
directly by the implementation.