remap button order to something more sensible
The hardware button ordering is based on where the traces could get from
buttons to the hc165. Create a virtual button ordering with a more human
friendly order. Reordered the throttle buttons to come before the wing
buttons in case programs support only a smaller number of buttons, since
the throttle buttons are in a more accessible location.